time_server/requirements.txt
django==2.2
requests==2.19.1
pyyaml==3.12
arrow==0.15.4
regex==2019.11.1